Database connection and docker setup
- Setup docker-compose
- Add .env config
- Connect postgres
- Add account repository

+++ internal/app/services/types.go
@@ -3,6 +3,6 @@ package services
 import "sum_changer_api/internal/app/models"
 
 type AccountService interface {
-	Get() (float64, error)
-	HandleOperation(role models.UserRole, sum float64) error
+	Get() (float32, error)
+	HandleOperation(role models.UserRole, sum float32) error
 }
